Added man-page for bbackupctl.
Makefile changes, to accommodate unlimited number of man-pages


git-svn-id: https://www.boxbackup.org/svn/box/trunk@555 9c52e268-1103-0410-960d-c6a436f6b474

+++ documentation/boxbackup/Makefile
@@ -4,7 +4,8 @@
 
 DBPROC=/usr/bin/xsltproc
 BOOKXSL=bb-book.xsl
-MANXSL= /usr/share/sgml/docbook/xsl-stylesheets/manpages/docbook.xsl
+NOCHUNKBOOKXSL=bb-nochunk-book.xsl
+MANXSL=bb-man.xsl
 VPATH= adminguide
 .SUFFIXES: .html .xml
 
@@ -24,25 +25,37 @@ instguide/index.html: instguide.xml $(BOOKXSL)
 ExceptionCodes.xml: ../../ExceptionCodes.txt
 	perl ./generate_except_xml.pl
 
-manpages: bbackupquery.html bbackupquery.1 
+manpages: man-dirs man-nroff man-html
 
-bbackupquery.html: bbackupquery.xml
-	$(DBPROC) -o man-html/bbackupquery.html $(BOOKXSL) bbackupquery.xml
+man-dirs: man-pages/.there man-html/.there
 
-bbackupquery.1: bbackupquery.xml
-	$(DBPROC) -o man-pages/bbackupquery.1 $(MANXSL) bbackupquery.xml
+man-html/.there:
+	if [ ! -d man-html ]; then mkdir man-html; touch man-html/.there; fi
 
-dockit: instguide adminguide
-	tar zcf documentation-kit-0.10.tar.gz html/ instguide/ adminguide/
+man-pages/.there:
+	if [ ! -d man-pages ]; then mkdir man-pages; touch man-pages/.there; fi
+
+man-nroff: bbackupquery.1 bbackupctl.1
+
+man-html: bbackupquery.html bbackupctl.html
+
+%.html: %.xml
+	$(DBPROC) -o $@ $(NOCHUNKBOOKXSL) $<
+	mv $@ man-html/.
+
+%.1: %.xml
+	$(DBPROC) -o $@ $(MANXSL) $<
+	mv $@ man-pages/.
+
+dockit: instguide adminguide manpages
+	tar zcf documentation-kit-0.10.tar.gz html/ instguide/ adminguide/ man-html/ man-pages
 
 clean:
-	rm -rf ./instguide/
-	rm -rf ./adminguide/
-	rm ExceptionCodes.xml
-	rm documentation-kit-0.10.tar.gz
-
-
-# %.html: %.xml
-# 	$(DBPROC) -o $@ $(BOOKXSL) $@.xml
+	if [ -d ./instguide ]; then rm -rf ./instguide/; fi
+	if [ -d ./adminguide ]; then rm -rf ./adminguide/; fi
+	if [ -d ./man-html ]; then rm -rf ./man-html/; fi
+	if [ -d ./man-pages ]; then  rm -rf ./man-pages/; fi
+	if [ -f ExceptionCodes.xml ]; then rm ExceptionCodes.xml; fi
+	if [ -f docuemtation-kit-0.10.tar.gz ]; then rm documentation-kit-0.10.tar.gz; fi
